+++ tests/ParserTest.php
@@ -55,7 +55,7 @@ class ParserTest extends TestCase
             'cats' => null,
             'name' => 'Elliot',
             'movie' => 'Mr. Robot',
-        ]);
+        ], ParserOption::PARSE_TEXT);
 
         $this->assertSame($expect, $parser->parseHtml());
     }